Add cart

+++ database/migrations/2023_07_19_121235_create_services_table.php
@@ -14,8 +14,8 @@ return new class extends Migration
         Schema::create('services', function (Blueprint $table) {
             $table->id();
             $table->string('name')->unique();
-            $table->decimal('price', 3, 2);
-            $table->integer('deadline');
+            $table->decimal('price', 10, 2);
+            $table->decimal('deadline', 10, 0);
             $table->timestamps();
         });
     }
